% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLoci.R
\name{drawLoci}
\alias{drawLoci}
\title{Draw alleles for loci list.}
\usage{
drawLoci(x)
}
\arguments{
\item{x}{List. Named list where names correspond to true loci names. Each element should hold unique allele values. At this time
they are assumed to be coercable to integers.}
}
\description{
Draw loci and its corresponding alleles.
}
\details{
This function will take a list of loci alleles and plot it, where y axis represents loci and x axis represents allele "length". Function
assumes allele names are coercable to numeric.
}
\examples{
# generate data
require(magrittr)
require(ggplot2)
al <- lapply(1:10, FUN = function(x) {
      runif(n = sample(x = 2:7, size = 1), min = 100, max = 300) \%>\%
      round()
  })

# add names to the list
names(al) <- paste("L", 1:length(al), sep = "")

fig.plotted <- drawLoci(al)
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})
}

