% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSeason.R
\name{findSeason}
\alias{findSeason}
\title{For a datum, find into which season it belongs to.}
\usage{
findSeason(x)
}
\arguments{
\item{x}{Datum of class Date.}
}
\value{
Character of the season.
}
\description{
Based on definition of seasons by solstices and equinoxes (quinoxi?), find to which a particular date belongs to. 
Function is based on Josh O'Brien's answer from StackOverflow 
(http://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to).
}
\examples{
findSeason(as.Date("21.6.2015", format = "\%d.\%m.\%Y"))
}
