% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genindSlidingWindow.R
\name{genindSlidingWindow}
\alias{genindSlidingWindow}
\title{Create a list of genind objects based on a sliding window.

Create a list of \code{genind} objects based on a sliding window of dates. Assumed step size is one year.
The resulting list can be written to GENEPOP file using \code{writeGenPop} function.}
\usage{
genindSlidingWindow(gi, start.year, win.size, dates, verbose = TRUE)
}
\arguments{
\item{gi}{A \code{genind} object to be subsetted.}

\item{start.year}{Integer. Starting year. See \code{win.size}.}

\item{win.size}{Integer. One side of sliding window. If \code{start.year} is 1995 and \code{win.size} is
2, sliding window will encompass years 1993:1997.}

\item{dates}{A vector of dates that correspond to individual genotype in \code{gi}. Often stored in \code{other(x)}.
Dates will be coerced to integer years (\%Y).}

\item{verbose}{Boolean. If TRUE, a message will be printed about cohort span. Can be silenced by setting this to FALSE.}
}
\value{
A list of \code{genind} objects.
}
\description{
Create a list of genind objects based on a sliding window.

Create a list of \code{genind} objects based on a sliding window of dates. Assumed step size is one year.
The resulting list can be written to GENEPOP file using \code{writeGenPop} function.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}

