% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/goSlide.R
\name{goSlide}
\alias{goSlide}
\title{Calculate sliding window cohorts}
\usage{
goSlide(start, win.size, cohorts)
}
\arguments{
\item{start}{Integer. Year of start of sliding window.}

\item{win.size}{Integer. Size of sliding window.}

\item{cohorts}{Integer. Number of cohorts (sliding window steps).}
}
\value{
A printout to console of cohort years.
}
\description{
In case you need to know years of cohorts, enter start date, sliding window size and number
of cohorts. Result is a visualization of from-to years of each cohort. This is a supporting
function to \code{genindSlidingWindow} function.
}
\examples{
goSlide(start = 1996, win.size = 4, cohorts = 14)
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})
}

