% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNGSfilter.R
\name{makeNGSfilter}
\alias{makeNGSfilter}
\title{Generate .ngsfilter file from a template}
\usage{
makeNGSfilter(sheet, template, experiment, num.replicates,
  sample.prefix = "sample", filename)
}
\arguments{
\item{sheet}{A data.frame with 8 rows and 12 columns. Each cell holds sample name. If file(path) is
provided, it assumes the data comes in an .xlsx file where the 8x12 frame starts in cell B4. 8x12 frame
should have numbers column names 1:12 from B3 to M3 and row names A:H from A4 to A11.}

\item{template}{A file (path) to the template where loci names, well position, tag and primers are defined.
See details for the layout.}

\item{experiment}{Character string denoting experiment name, e.g. UA_FrSlo. First two letters should be
species initials, followed by an underscore and followed by a project name. Locus name will be added to
this name, e.g. UA_FrSlo_03. The result should have exactly two underscores "_".}

\item{num.replicates}{Integer. Number of replicates.}

\item{filename}{Name (path) of the file to be created.}
}
\value{
A data.frame and a side effect of creating the file.
}
\description{
.ngsfilter file holds sample names and maps it to the appropriate tag-primer combination for each well on 
plate(s). Tag-primer combination for each well is stored in a template user provides. This template is 
appended by an experiment and sample names from "sheet" and accounted for the number of replicates.
}
\details{
`template` should have seven (7) columns, as described below. Notice the number of underscores.
\itemize{
   \item Experiment name, followed by locus name, e.g. \code{UA_FrSlo_03} where \code{UA_FrSlo} is name
   of the experiment and \code{03} is locus name.
   \item Sample name, well number and replicate, e.g. UASwd_001_R1 where UASwd is sample name, 001 is the
   first well A1 and R1 means replicate 1.
   \item Tag combination for given forward/reverse primer combination, e.g. \code{acacacac:acacacac}.
   \item Forward primer, e.g. \code{gctcccataactgcataaggtc}.
   \item Reverse primer, e.g. \code{ctggctggctggctagg}.
   \item Column for comments. We keep all values as \code{F}.
   \item Column where we keep all values as \code{@}.
}

This function is the result of our visit to Grenoble (France) in collaboration with Marta de Barba and her 
colleagues (Laboratoire d’Écologie Alpine). Data is from paper by de Barba et al. (2016) and is directly 
suitable for the protocol described in the Supplemental materials of the above paper.

De Barba, M., Miquel, C., Lobréaux, S., Quenette, P. Y., Swenson, J. E., & Taberlet, P. (2016). 
High-throughput microsatellite genotyping in ecology: improved accuracy, efficiency, standardization 
and success with low-quantity and degraded DNA. Molecular Ecology Resources, 1–16. 
http://doi.org/10.1111/1755-0998.12594
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}

