% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mratio.R
\name{Mratio}
\alias{Mratio}
\title{Calculate the Garza-Williamson M-ratio.}
\usage{
Mratio(x)
}
\arguments{
\item{x}{A \code{genind} object.}
}
\value{
A named vector of loci with M ratio.
}
\description{
M-ratio indicates reduction in population size using data from microsatellite loci.
}
\details{
Formula for calculating for each locus is M = k/(R+1) where \code{k} is the number of alleles on that locus
and R is the range of microsatellite repeats. See Garza J.C., Williamson E.G. 2001. Detection of reduction 
in population size using data from microsatellite loci. Molecular Ecology, 10: 305-318 for more information.
}
\examples{
library(adegenet)
data(nancycats)
Mratio(nancycats)
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
