% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcrplate}
\alias{pcrplate}
\title{PCR plate sample name layout}
\format{A data.frame with 8 rows and 12 columns. Each well contains sample name.}
\source{
Laboratory at Department of biology, Biotechnical faculty, University of Ljubljana.
}
\usage{
pcrplate
}
\description{
8x12 PCR plate with sample names for each well.
}
\keyword{datasets}

