% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCircleOnMap.R
\name{plotCircleOnMap}
\alias{plotCircleOnMap}
\title{Plot circles on a map}
\usage{
plotCircleOnMap(coords, multiply.r, multiply.se, map)
}
\arguments{
\item{coords}{\code{data.frame} with columns latitude, longitude and projection already defined as in \code{sp}.}

\item{multiply.r}{Numeric. Multiplicator that will inflate or deflate circle (original in miles).}

\item{multiply.se}{Numeric. Same as for \code{multiply.r} but for line width (standard error).}

\item{map}{A \code{RasterLayer} object from \code{dismo:::gmap}.}
}
\value{
A list of input data and map used in plotting.
}
\description{
This is not yet a fully general function for plotting circles on a google map as presented by the \code{dismo} package.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}

