% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readColony.R
\name{readColony}
\alias{readColony}
\title{Function will read in result from Colony.}
\usage{
readColony(proj, comp = c("LC", "LR", "offspring_genotype"))
}
\arguments{
\item{proj}{Character. A full path to the Colony project folder.}

\item{comp}{Character. Result from colony. See Details.}
}
\description{
Specifying the project folder, function will find the desired result and import it into a
R specific data structure.
}
\section{Details}{

\itemize{
\item \code{"LC"} will extract result from \code{*.BestClone}.
\item \code{"LR"} (not implemented yet) will extract results from \code{*.PairwiseCloneDyad}. 
This will be empty if relatedness wasn't calculated.
\item \code{offspring_genotype} will extract genotype of offspring from \code{OffspringGenotype.txt} of the project.
}
}

\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})
}
