% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampleGenData.R
\name{subsetGenData}
\alias{subsetGenData}
\title{Handles dropping levels of subsetted data}
\usage{
subsetGenData(x)
}
\arguments{
\item{x}{A subsetted \code{genind} object, possibly with other fields in \code{other} slot.}
}
\description{
Get a fully subsetted \code{genind} object. \code{genind} methods handles subsetting and this
functions drops levels by reconstructing the object from scratch.
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}
