% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/throwoutKin.R
\name{throwoutKin}
\alias{throwoutKin}
\title{Function finds non related individuals.}
\usage{
throwoutKin(data, treshold)
}
\arguments{
\item{data}{\code{data.frame} A \code{data.frame} with columns (in this order) \code{animal1},
\code{animal2} and \code{stat}.}

\item{treshold}{\code{numeric} A numeric vector of length one. Individuals with comparison
value of >= \code{treshold} will be removed from the data set.}
}
\description{
For each individual, remove kin based on treshold value (see Blouin 2003). Data comes from
software \code{coancestry} (Jinliang Wang).
}
\examples{
\dontrun{
  kin <- read.table("./sandbox/RelatednessCI95.Txt", sep = ",")
  animal <- kin[, c(2, 3, 6)]
  names(animal) <- c("animal1", "animal2", "stat")

  suppressMessages(rez <- throwoutKin(data = animal, treshold = 0.5))
  str(rez)
  length(rez$unique.animals) # number of unique animals
  str(rez$cleaned.data) # structure of cleaned data
  hist(rez$cleaned.data$stat) # histogram of stat values from cleaned data
                              # should all be below treshold
}
}
\author{
Roman Lustrik, Maja Jelencic
}
