% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeCoancestry.R
\name{writeCoancestry}
\alias{writeCoancestry}
\title{Function will write \code{genind} object to Coancestry file format.}
\usage{
writeCoancestry(gi, file)
}
\arguments{
\item{gi}{A \code{genind} object to be written to a file.}

\item{file}{Character. A file name of the created file.}
}
\value{
A file is written to disk. Rows (individuals) are sorted according to the population quietly.
}
\description{
Using \code{genind} object, function will try hard to format and write the data
so that it will be accepted by Coancestry program (http://www.zsl.org/science/software/coancestry).
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})
}

