% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeGenePop}
\alias{writeGenePop}
\title{Write genind object to text files}
\usage{
writeGenePop(genotypes1, genotypes2, path, note)
}
\arguments{
\item{genotypes1}{genind A \code{genind} object.}

\item{genotypes2}{genind A \code{genind} object}

\item{path}{Character A relative or absolute path where file is written to.}

\item{note}{Character A character vector of length 1 reserved for notes.}
}
\value{
Text file
}
\description{
Function promises to write two genind objects to a text file.
}
\author{
Tomaz Skrbinsek {tomaz.skrbinsek@gmail.com}
}

