% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/writeGenPop.R
\name{writeGenPop}
\alias{writeGenPop}
\title{Write genind as GENEPOP file}
\usage{
writeGenPop(gi, file.name, comment)
}
\arguments{
\item{gi}{A \code{\link[adegenet]{genind}} object with defined strata and population. See \code{\link[adegenet]{setPop}}.}

\item{file.name}{A character string defining file name. You will have to specify the extension as well.}

\item{comment}{A character string with a desired comment that will be prepended to the beginning of the file.}
}
\value{
A file on disk, function return NULL to R.
}
\description{
Write \code{adegenet}'s \code{genind} object to GENEPOP format
}
\details{
Function will write \code{adegenet}'s \code{genind} object to GENEPOP as defined at
\url{http://genepop.curtin.edu.au/help_input.html}. First line is a comment, second line is loci names and the rest consists
of lines "pop" and genotype names and their allele values on corresponding loci. This function differs from
\code{\link{writeGenePop}}, relies on defined \code{strata} and is more general.
}
\examples{
library(adegenet)
data(nancycats)
# Genind object must have a population defined. This is trivial with strata.
other(nancycats)$pop <- sample(letters[1:4], size = nInd(nancycats), replace = TRUE)
strata(nancycats) <- data.frame(pop = other(nancycats)$pop)
setPop(nancycats) <- ~pop

# write the file and remove it
writeGenPop(nancycats, file.name = "test.gen", comment = "is a no comment")
file.remove("test.gen")
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})
}

