% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeINEST}
\alias{writeINEST}
\title{Write genind object as INEST file}
\usage{
writeINEST(x, file, spatial = FALSE)
}
\arguments{
\item{x}{\code{genind} object}

\item{file}{Character vector. Path where file is written to.}

\item{spatial}{Logical. If TRUE, spatial coordinates of samples will be written to file.}
}
\description{
Write genind object as INEST file
}
\examples{
library(adegenet)
data(nancycats)
writeINEST(nancycats, file = "inest_test.txt", spatial = FALSE)
unlink("inest_test.txt")
}
\author{
Roman Lustrik (roman.lustrik@biolitika.si)
}

