% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeStructure}
\alias{writeStructure}
\title{Function writes a genind object to a text file}
\usage{
writeStructure(genotipi, pop, path)
}
\arguments{
\item{pop}{Vector A numeric vector which corresponds to population.}

\item{path}{Character A character vector of relative or absolute path of the resulting file.}

\item{genotypes}{genind A \code{genind} object to be exported.}
}
\value{
Text file
}
\description{
Function will convert a genind file and population designation to a text file.
}
\author{
Tomaz Skrbinsek {tomaz.skrbinsek@gmail.com}
}

