% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{zvau-package}
\alias{zvau}
\alias{zvau-package}
\title{Misc functions used by out group}
\description{
Package contains functions used in daily work by our group for ecology at
Department of biology, Biotechnical faculty, University of Ljubljana.
}
\details{
So far the functions included in the package are:
\itemize{
  \item \code{\link{read.clumpp}} Read result of Clumpp
}

\tabular{ll}{
Package: \tab zvau\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-31\cr
License: \tab GPL-2\cr
LazyData: \tab no\cr
}
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si})

Maintainer: Roman Lustrik (\email{roman.lustrik@biolitika.si})
}
\keyword{datasets}
\keyword{package}

