% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeStructure.R
\name{writeStructure}
\alias{writeStructure}
\title{Function writes a genind object to a text file}
\usage{
writeStructure(x, file)
}
\arguments{
\item{x}{genind A \code{genind} object to be exported.}

\item{file}{Character A character vector of relative or absolute path of the resulting file.}
}
\value{
Writese a text file to a specified path or current \code{getwd()}. Returns
a population legend if population is designated. Otherwise returns a message notifying
the user legend is not applicable.
}
\description{
Function will convert a genind file and population designation to a text file.
}
\author{
Tomaz Skrbinsek (\email{tomaz.skrbinsek@gmail.com}) and Roman Lustrik (\email{roman.lustrik@biolitika.si})
}
