% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/writeStructure.R
\name{writeStructure}
\alias{writeStructure}
\title{Function writes a genind object to a text file}
\usage{
writeStructure(genotipi, pop, path, ...)
}
\arguments{
\item{genotipi}{genind A \code{genind} object to be exported.}

\item{pop}{Vector A numeric vector which corresponds to population.}

\item{path}{Character A character vector of relative or absolute path of the resulting file.}

\item{...}{Arguments passed to \code{write.table}.}
}
\value{
Text file
}
\description{
Function will convert a genind file and population designation to a text file.
}
\author{
Tomaz Skrbinsek (\email{tomaz.skrbinsek@gmail.com})
}

