% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zvau-package.R
\docType{package}
\name{zvau-package}
\alias{zvau}
\alias{zvau-package}
\title{Misc functions used by out group}
\description{
Package contains functions used in daily work by our group for ecology at
Department of biology, Biotechnical faculty, University of Ljubljana.
}
\details{
So far the functions included in the package are:
\itemize{
  \item \code{\link{readClumpp}} Read result of Clumpp into a \code{data.frame}
  \item \code{\link{writeStructure}} Write a \code{genind} object to be analyzed using Structure
  \item \code{\link{writeGenePop}} Write a \code{genind} object to be analyzed using GENEPOP
  \item \code{\link{drawLoci}} Plot alleles using a list of loci
}

\tabular{ll}{
Package: \tab zvau\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2015-02-11\cr
License: \tab GPL-2\cr
LazyData: \tab no\cr
}
}
\author{
Roman Lustrik (\email{roman.lustrik@biolitika.si}), Tomaz Skrbinsek ( \email{tomaz.skrbinsek@gmail.com}), Maja Jelencic (\email{mjelencic@gmail.com})

Maintainer: Roman Lustrik (\email{roman.lustrik@biolitika.si})
}
\keyword{datasets}
\keyword{package}

